Config : Adding support for Immutant config
Retaining support for Jetty and working in REPL
via tools.namespace config.

Note: To allow for the project config to support
running via the REPL I needed to exclude the
immutant/init.clj from being reloaded at the REPL
for a couple of reasons:
- init.clj has dependencies on libraries in the
  immutant runtime and I did not want to add those
  as general dependencies.
- the tools.namespace refresh resulted in an error
  about the immutant messaging service bein started
  outside the context of an Immutant instance.

To control the dirs for which the source files get loaded
by tools.namespace, I used the tools.namespace/set-refresh-dirs
function in the reset function in user.clj.

This may not be the best approach, but for now, this lets us
incorporate Immutant support without impacting the way the
VLACS projects have been developed up to this point.

--Tom